requirements/base.txt
celery

coverage
Django
django-environ
django-extensions
django-cors-headers
Pillow
pre-commit
psycopg2-binary
requests
wget

numpy
matplotlib